+++ routes/web.php
@@ -164,6 +164,7 @@ Route::prefix('errors')->name('errors.')->group(function () {
 
 Route::get('/mock-api/flights', [\App\Http\Controllers\MockApi\OpenSkyController::class, 'index'])->name('mock-api.flights');
 Route::get('/mock-api/satellites', [\App\Http\Controllers\MockApi\CelesTrakController::class, 'index'])->name('mock-api.satellites');
+Route::get('/mock-api/places', [\App\Http\Controllers\MockApi\OverpassController::class, 'index'])->name('mock-api.places');
 Route::post('/mock-api/google-maps/session', [\App\Http\Controllers\MockApi\GoogleMapsController::class, 'createSession'])->name('mock-api.gmaps.session');
 Route::get('/mock-api/google-maps/config', [\App\Http\Controllers\MockApi\GoogleMapsController::class, 'config'])->name('mock-api.gmaps.config');